rename auth_view to user_view
